% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_sellers_fiction.R
\name{best_sellers_fiction}
\alias{best_sellers_fiction}
\title{Title}
\usage{
best_sellers_fiction()
}
\value{
Returns books that have fallen in rank on newyork times best seller list from hardcover-fiction
}
\description{
Title
}
\examples{
best_sellers_fiction()
}
