% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJIdata.R
\docType{data}
\name{DJIdata}
\alias{DJIdata}
\title{Dow Jones Industrial Average}
\format{
241 * 7 dataframe

\describe{
\item{Index}{date}
\item{Variables}{Open}{High}{Low}{Close}{Adj Close}{Volume}
}
}
\usage{
DJIdata
}
\description{
Dow Jones Industrial Average
}
\keyword{datasets}
