% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_columns.R
\name{drop_columns}
\alias{drop_columns}
\title{Drop colums}
\usage{
drop_columns(df, threshold)
}
\arguments{
\item{df}{This is a structured dataframe}

\item{threshold}{Cutoff for missing data from 0-1}
}
\value{
return a new data frame with dropped NAs
}
\description{
droping variables that have too many NA values
}
