% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDPC1.R
\docType{data}
\name{GDPC1}
\alias{GDPC1}
\title{GDPC1}
\format{
295 x 2 dataframe

\describe{
\item{index}{Date YY-MM-DD}
\item{column1}{Real Gross Domestic Product in Billions }}
}
\usage{
GDPC1
}
\description{
GDPC1
}
\keyword{datasets}
