% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HealthData.R
\docType{data}
\name{HealthData}
\alias{HealthData}
\title{some World Bank Data}
\format{
264 x 64 dataframe ...

\describe{
\item{index}{Country Names}
\item{column1}{Literacy rates}
}
}
\usage{
HealthData
}
\description{
some World Bank Data
}
\keyword{datasets}
