% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Life_Expectancy_Data.R
\docType{data}
\name{Life_Expectancy_Data}
\alias{Life_Expectancy_Data}
\title{Life_Expectancy_Data}
\format{
2938 x 22 dataframe
}
\usage{
Life_Expectancy_Data
}
\description{
Life_Expectancy_Data
}
\keyword{datasets}
