% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_num.R
\name{na_num}
\alias{na_num}
\title{na_num}
\usage{
na_num(df)
}
\arguments{
\item{df}{Given data frame}
}
\value{
the number of NA in this given data frame
}
\description{
to check the number of NA in the structure data frame
}
