% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/plus.R
\name{plus}
\alias{plus}
\title{A+B}
\usage{
plus(a, b, c)

plus(a, b, c)
}
\arguments{
\item{a}{interger}

\item{b}{interger}
}
\value{
a+b


}
\description{
plus both variablllell
}
\examples{
5+4+2 = 11
}
