% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_quadratic.R
\name{seq_quadratic}
\alias{seq_quadratic}
\title{seq_quadratic}
\usage{
seq_quadratic(x)
}
\arguments{
\item{x}{a number}
}
\value{

}
\description{
input a number, print a sequence of input's quadratic from 1 to the input value
}
