% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Square All.R
\name{square_all}
\alias{square_all}
\title{Square Loop}
\usage{
square_all(x)
}
\arguments{
\item{x}{Any integer you want to square, from 1 to x}
}
\value{

}
\description{
Square Loop
}
\examples{
square_all(10)
}
