% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNHdata.R
\docType{data}
\name{UNHdata}
\alias{UNHdata}
\title{stock price of UnitedHealth Group Incorporated}
\format{
241 * 7 dataframe

\describe{
\item{Index}{date}
\item{Variables}{Open}{High}{Low}{Close}{Adj Close}{Volume}
}
}
\usage{
UNHdata
}
\description{
stock price of UnitedHealth Group Incorporated
}
\keyword{datasets}
