% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_columns.R
\name{drop_columns}
\alias{drop_columns}
\title{Title}
\usage{
drop_columns(df, threshold)
}
\arguments{
\item{df}{This is a dataframe with any datatypes}

\item{threshold}{Between 0-1 for the amount of NAs we accept in a column}
}
\value{
Returns a df w droppe dcolumsn
}
\description{
Title
}
\examples{
drop_columns(df, .1)
}
