% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop columns.R
\name{drop_columns}
\alias{drop_columns}
\title{Drops NA}
\usage{
drop_columns(df, threshold)
}
\arguments{
\item{df}{Any dataframe}

\item{threshold}{Cutoff for missing data from 0 to 1}
}
\value{

}
\description{
Drops NA
}
\examples{
drop_columns(df, .2)
}
